Refactor build and package scripts, and improve cleanup

+++ repack.ps1
@@ -8,8 +8,17 @@ dotnet build
 dotnet pack  /p:PackageVersion=2.0.0
 
 # copy the dotnet-interactive packages to the temp directory
-Get-ChildItem -Recurse -Filter *.nupkg | Move-Item -Destination c:\temp\packages -Force
+$destinationPath = "C:\temp\packages"
+if (-not (Test-Path -Path $destinationPath -PathType Container)) {
+    New-Item -Path $destinationPath -ItemType Directory -Force
+}
+Get-ChildItem -Recurse -Filter *.nupkg | Move-Item -Destination $destinationPath -Force
 
 # delete the #r nuget caches
-Remove-Item -Recurse -Force ~\.packagemanagement\nuget\Cache
-Remove-Item -Recurse -Force ~\.packagemanagement\nuget\Projects
+if (Test-Path -Path ~\.packagemanagement\nuget\Cache -PathType Container) {
+    Remove-Item -Recurse -Force ~\.packagemanagement\nuget\Cache
+}
+
+if (Test-Path -Path ~\.packagemanagement\nuget\Projects -PathType Container) {
+    Remove-Item -Recurse -Force ~\.packagemanagement\nuget\Projects
+}
